Make all coordinates 16bit integers to allow for bigger matrices (max 256x256 pixels) without losing accurate particle positioning (PS_P_RADIUS = 32) Added FastLEDPartMatrix to render the particles directly into leds array. Depends on FastLED library ( fastled.io )

+++ ParticleSys.cpp
@@ -26,7 +26,7 @@ void ParticleSys::update()
 {
     cycleRemaining = perCycle;
     emitter->update();
-    for(int i = 0; i<num; i++) {
+    for(byte i=0; i<num; i++) {
         if (!particles[i].isAlive && cycleRemaining > 0) {
             emitter->emit(&particles[i]);
             cycleRemaining--;
